+++ crates/client/rpc/src/lib.rs
@@ -6,11 +6,15 @@ mod constants;
 mod errors;
 mod events;
 mod madara_backend_client;
+mod runtime_api;
 mod trace_api;
 mod types;
+
+use std::collections::HashMap;
 use std::marker::PhantomData;
 use std::sync::Arc;
 
+use blockifier::transaction::objects::ResourcesMapping;
 use errors::StarknetRpcApiError;
 use jsonrpsee::core::{async_trait, RpcResult};
 use jsonrpsee::types::error::CallError;
@@ -33,7 +37,7 @@ use sc_client_api::BlockBackend;
 use sc_network_sync::SyncingService;
 use sc_transaction_pool::{ChainApi, Pool};
 use sc_transaction_pool_api::error::{Error as PoolError, IntoPoolError};
-use sc_transaction_pool_api::{TransactionPool, TransactionSource};
+use sc_transaction_pool_api::{InPoolTransaction, TransactionPool, TransactionSource};
 use sp_api::ProvideRuntimeApi;
 use sp_arithmetic::traits::UniqueSaturatedInto;
 use sp_blockchain::HeaderBackend;
@@ -43,16 +47,18 @@ use sp_runtime::transaction_validity::InvalidTransaction;
 use sp_runtime::DispatchError;
 use starknet_api::block::BlockHash;
 use starknet_api::hash::StarkHash;
-use starknet_api::transaction::Calldata;
+use starknet_api::transaction::{Calldata, TransactionHash};
 use starknet_core::types::{
     BlockHashAndNumber, BlockId, BlockStatus, BlockTag, BlockWithTxHashes, BlockWithTxs, BroadcastedDeclareTransaction,
     BroadcastedDeployAccountTransaction, BroadcastedInvokeTransaction, BroadcastedTransaction, ContractClass,
     DeclareTransactionReceipt, DeclareTransactionResult, DeployAccountTransactionReceipt,
     DeployAccountTransactionResult, EventFilterWithPage, EventsPage, ExecutionResources, ExecutionResult, FeeEstimate,
     FieldElement, FunctionCall, Hash256, InvokeTransactionReceipt, InvokeTransactionResult,
-    L1HandlerTransactionReceipt, MaybePendingBlockWithTxHashes, MaybePendingBlockWithTxs,
-    MaybePendingTransactionReceipt, MsgFromL1, StateDiff, StateUpdate, SyncStatus, SyncStatusType, Transaction,
-    TransactionExecutionStatus, TransactionFinalityStatus, TransactionReceipt,
+    L1HandlerTransactionReceipt, MaybePendingBlockWithTxHashes, MaybePendingBlockWithTxs, MaybePendingStateUpdate,
+    MaybePendingTransactionReceipt, MsgFromL1, PendingBlockWithTxHashes, PendingBlockWithTxs,
+    PendingDeclareTransactionReceipt, PendingDeployAccountTransactionReceipt, PendingInvokeTransactionReceipt,
+    PendingL1HandlerTransactionReceipt, PendingStateUpdate, PendingTransactionReceipt, StateDiff, StateUpdate,
+    SyncStatus, SyncStatusType, Transaction, TransactionExecutionStatus, TransactionFinalityStatus, TransactionReceipt,
 };
 use starknet_core::utils::get_selector_from_name;
 
@@ -319,7 +325,7 @@ where
         &self,
         declare_transaction: BroadcastedDeclareTransaction,
     ) -> RpcResult<DeclareTransactionResult> {
-        let best_block_hash = self.client.info().best_hash;
+        let best_block_hash = self.get_best_block_hash();
 
         let opt_sierra_contract_class = if let BroadcastedDeclareTransaction::V2(ref tx) = declare_transaction {
             Some(flattened_sierra_to_sierra_contract_class(tx.contract_class.clone()))
@@ -336,7 +342,7 @@ where
             _ => Err(StarknetRpcApiError::InternalServerError)?,
         };
 
-        let current_block_hash = self.client.info().best_hash;
+        let current_block_hash = self.get_best_block_hash();
         let contract_class = self
             .overrides
             .for_block_hash(self.client.as_ref(), current_block_hash)
@@ -347,7 +353,7 @@ where
             return Err(StarknetRpcApiError::ClassAlreadyDeclared.into());
         }
 
-        let extrinsic = convert_tx_to_extrinsic(self.client.clone(), best_block_hash, transaction.clone()).await?;
+        let extrinsic = self.convert_tx_to_extrinsic(best_block_hash, transaction.clone())?;
 
         submit_extrinsic(self.pool.clone(), best_block_hash, extrinsic).await?;
 
@@ -382,14 +388,14 @@ where
         &self,
         invoke_transaction: BroadcastedInvokeTransaction,
     ) -> RpcResult<InvokeTransactionResult> {
-        let best_block_hash = self.client.info().best_hash;
+        let best_block_hash = self.get_best_block_hash();
 
         let transaction: UserTransaction = invoke_transaction.try_into().map_err(|e| {
             error!("Failed to convert BroadcastedInvokeTransaction to UserTransaction: {e}");
             StarknetRpcApiError::InternalServerError
         })?;
 
-        let extrinsic = convert_tx_to_extrinsic(self.client.clone(), best_block_hash, transaction.clone()).await?;
+        let extrinsic = self.convert_tx_to_extrinsic(best_block_hash, transaction.clone())?;
 
         submit_extrinsic(self.pool.clone(), best_block_hash, extrinsic).await?;
 
@@ -412,14 +418,14 @@ where
         &self,
         deploy_account_transaction: BroadcastedDeployAccountTransaction,
     ) -> RpcResult<DeployAccountTransactionResult> {
-        let best_block_hash = self.client.info().best_hash;
+        let best_block_hash = self.get_best_block_hash();
 
         let transaction: UserTransaction = deploy_account_transaction.try_into().map_err(|e| {
             error!("Failed to convert BroadcastedDeployAccountTransaction to UserTransaction, error: {e}",);
             StarknetRpcApiError::InternalServerError
         })?;
 
-        let extrinsic = convert_tx_to_extrinsic(self.client.clone(), best_block_hash, transaction.clone()).await?;
+        let extrinsic = self.convert_tx_to_extrinsic(best_block_hash, transaction.clone())?;
 
         submit_extrinsic(self.pool.clone(), best_block_hash, extrinsic).await?;
 
@@ -578,17 +584,7 @@ where
             };
 
         let execution_status = {
-            let revert_error = self
-                .client
-                .runtime_api()
-                .get_tx_execution_outcome(substrate_block_hash, Felt252Wrapper(transaction_hash).into())
-                .map_err(|e| {
-                    error!(
-                        "Failed to get transaction execution outcome. Substrate block hash: {substrate_block_hash}, \
-                         transaction hash: {transaction_hash}, error: {e}"
-                    );
-                    StarknetRpcApiError::InternalServerError
-                })?;
+            let revert_error = self.get_tx_execution_outcome(substrate_block_hash, transaction_hash)?;
 
             if revert_error.is_none() {
                 TransactionExecutionStatus::Succeeded
@@ -677,23 +673,16 @@ where
             StarknetRpcApiError::BlockNotFound
         })?;
 
-        let runtime_api = self.client.runtime_api();
-
         let calldata = Calldata(Arc::new(request.calldata.iter().map(|x| Felt252Wrapper::from(*x).into()).collect()));
 
-        let result = runtime_api
-            .call(
-                substrate_block_hash,
-                Felt252Wrapper(request.contract_address).into(),
-                Felt252Wrapper(request.entry_point_selector).into(),
-                calldata,
-            )
-            .map_err(|e| {
-                error!("Request parameters error: {e}");
-                StarknetRpcApiError::InternalServerError
-            })?;
+        let result = self.do_call(
+            substrate_block_hash,
+            Felt252Wrapper(request.contract_address).into(),
+            Felt252Wrapper(request.entry_point_selector).into(),
+            calldata,
+        )?;
 
-        let result = convert_error(self.client.clone(), substrate_block_hash, result)?;
+        let result = self.convert_error(substrate_block_hash, result)?;
 
         Ok(result.iter().map(|x| format!("{:#x}", x.0)).collect())
     }
@@ -890,14 +879,19 @@ where
     /// a pending block with transaction hashes, depending on the state of the requested block.
     /// In case the block is not found, returns a `StarknetRpcApiError` with `BlockNotFound`.
     fn get_block_with_tx_hashes(&self, block_id: BlockId) -> RpcResult<MaybePendingBlockWithTxHashes> {
+        let chain_id = self.chain_id()?;
+
+        if is_pending_block(block_id) {
+            let pending_block = self.prepare_pending_block_with_tx_hashes(Felt252Wrapper(chain_id.0))?;
+            return Ok(MaybePendingBlockWithTxHashes::PendingBlock(pending_block));
+        }
+
         let substrate_block_hash = self.substrate_block_hash_from_starknet_block(block_id).map_err(|e| {
             error!("'{e}'");
             StarknetRpcApiError::BlockNotFound
         })?;
 
         let starknet_block = get_block_by_block_hash(self.client.as_ref(), substrate_block_hash)?;
-
-        let chain_id = self.chain_id()?;
         let starknet_version = starknet_block.header().protocol_version;
         let l1_gas_price = starknet_block.header().l1_gas_price;
         let block_hash = starknet_block.header().hash::<H>();
@@ -993,11 +987,8 @@ where
     /// Returns the chain id this node is connected to. The chain id is returned as a specific type,
     /// defined by the Starknet protocol, indicating the particular network.
     fn chain_id(&self) -> RpcResult<Felt> {
-        let best_block_hash = self.client.info().best_hash;
-        let chain_id = self.client.runtime_api().chain_id(best_block_hash).map_err(|e| {
-            error!("Failed to fetch chain_id with best_block_hash: {best_block_hash}, error: {e}");
-            StarknetRpcApiError::InternalServerError
-        })?;
+        let best_block_hash = self.get_best_block_hash();
+        let chain_id = self.get_chain_id(best_block_hash)?;
 
         Ok(Felt(chain_id.0))
     }
@@ -1021,7 +1012,7 @@ where
             error!("'{e}'");
             StarknetRpcApiError::BlockNotFound
         })?;
-        let best_block_hash = self.client.info().best_hash;
+        let best_block_hash = self.get_best_block_hash();
         let chain_id = Felt252Wrapper(self.chain_id()?.0);
 
         let transactions =
@@ -1030,18 +1021,7 @@ where
                 StarknetRpcApiError::InternalServerError
             })?;
 
-        let fee_estimates = self
-            .client
-            .runtime_api()
-            .estimate_fee(substrate_block_hash, transactions)
-            .map_err(|e| {
-                error!("Request parameters error: {e}");
-                StarknetRpcApiError::InternalServerError
-            })?
-            .map_err(|e| {
-                error!("Failed to call function: {:#?}", e);
-                StarknetRpcApiError::ContractError
-            })?;
+        let fee_estimates = self.estimate_fee(best_block_hash, transactions)?;
 
         let estimates = fee_estimates
             .into_iter()
@@ -1080,18 +1060,7 @@ where
             StarknetRpcApiError::InternalServerError
         })?;
 
-        let fee_estimate = self
-            .client
-            .runtime_api()
-            .estimate_message_fee(substrate_block_hash, message)
-            .map_err(|e| {
-                error!("Runtime api error: {e}");
-                StarknetRpcApiError::InternalServerError
-            })?
-            .map_err(|e| {
-                error!("function execution failed: {:#?}", e);
-                StarknetRpcApiError::ContractError
-            })?;
+        let fee_estimate = self.do_estimate_message_fee(substrate_block_hash, message)?;
 
         let estimate = FeeEstimate {
             gas_price: fee_estimate.0.try_into().map_err(|_| StarknetRpcApiError::InternalServerError)?,
@@ -1171,6 +1140,14 @@ where
     /// transactions. In case the specified block is not found, returns a `StarknetRpcApiError` with
     /// `BlockNotFound`.
     fn get_block_with_txs(&self, block_id: BlockId) -> RpcResult<MaybePendingBlockWithTxs> {
+        let chain_id = self.chain_id()?;
+        let chain_id = Felt252Wrapper(chain_id.0);
+
+        if is_pending_block(block_id) {
+            let pending_block = self.prepare_pending_block_with_txs(chain_id)?;
+            return Ok(MaybePendingBlockWithTxs::PendingBlock(pending_block));
+        }
+
         let substrate_block_hash = self.substrate_block_hash_from_starknet_block(block_id).map_err(|e| {
             error!("Block not found: '{e}'");
             StarknetRpcApiError::BlockNotFound
@@ -1180,9 +1157,6 @@ where
 
         let block_hash = starknet_block.header().hash::<H>();
         let starknet_version = starknet_block.header().protocol_version;
-
-        let chain_id = self.chain_id()?;
-        let chain_id = Felt252Wrapper(chain_id.0);
 
         let opt_cached_transaction_hashes =
             self.get_cached_transaction_hashes(starknet_block.header().hash::<H>().into());
@@ -1237,7 +1211,25 @@ where
     /// the state of the network as a result of the block's execution. This can include a confirmed
     /// state update or a pending state update. If the block is not found, returns a
     /// `StarknetRpcApiError` with `BlockNotFound`.
-    fn get_state_update(&self, block_id: BlockId) -> RpcResult<StateUpdate> {
+    fn get_state_update(&self, block_id: BlockId) -> RpcResult<MaybePendingStateUpdate> {
+        if is_pending_block(block_id) {
+            let state_diff = StateDiff {
+                storage_diffs: Vec::new(),
+                deprecated_declared_classes: Vec::new(),
+                declared_classes: Vec::new(),
+                deployed_contracts: Vec::new(),
+                replaced_classes: Vec::new(),
+                nonces: Vec::new(),
+            };
+
+            let latest_block = self.get_best_block_hash();
+            let latest_block = get_block_by_block_hash(self.client.as_ref(), latest_block).unwrap_or_default();
+            let old_root = self.backend.temporary_global_state_root_getter().into();
+            let pending_state_update = PendingStateUpdate { old_root, state_diff };
+
+            return Ok(MaybePendingStateUpdate::PendingUpdate(pending_state_update));
+        }
+
         let substrate_block_hash = self.substrate_block_hash_from_starknet_block(block_id).map_err(|e| {
             error!("'{e}'");
             StarknetRpcApiError::BlockNotFound
@@ -1267,12 +1259,14 @@ where
             StarknetRpcApiError::InternalServerError
         })?;
 
-        Ok(StateUpdate {
+        let state_update = StateUpdate {
             block_hash: starknet_block.header().hash::<H>().into(),
             new_root: Felt252Wrapper::from(self.backend.temporary_global_state_root_getter()).into(),
             old_root,
             state_diff,
-        })
+        };
+
+        Ok(MaybePendingStateUpdate::Update(state_update))
     }
 
     /// Returns all events matching the given filter.
@@ -1433,18 +1427,107 @@ where
         &self,
         transaction_hash: FieldElement,
     ) -> RpcResult<MaybePendingTransactionReceipt> {
-        let substrate_block_hash = self
-            .backend
-            .mapping()
-            .block_hash_from_transaction_hash(transaction_hash.into())
-            .map_err(|e| {
+        let chain_id = self.chain_id()?.0.into();
+        let receipt =
+            match self.backend.mapping().block_hash_from_transaction_hash(transaction_hash.into()).map_err(|e| {
                 error!("Failed to interact with db backend error: {e}");
                 StarknetRpcApiError::InternalServerError
-            })?
-            .ok_or(StarknetRpcApiError::TxnHashNotFound)?;
+            })? {
+                Some(substrate_block_hash) => {
+                    self.prepare_tx_receipt(chain_id, transaction_hash, substrate_block_hash).await?
+                }
+                // Try to find pending Tx
+                None => self.get_pending_transaction_receipt(chain_id, transaction_hash).await.map_err(|e| {
+                    error!("Failed to find pending tx with hash: {transaction_hash}: {e}");
+                    StarknetRpcApiError::TxnHashNotFound
+                })?,
+            };
+        Ok(receipt)
+    }
+}
 
-        let starknet_block: mp_block::Block = get_block_by_block_hash(self.client.as_ref(), substrate_block_hash)?;
+/// RPC Helper methods
+impl<A, B, BE, G, C, P, H> Starknet<A, B, BE, G, C, P, H>
+where
+    A: ChainApi<Block = B> + 'static,
+    B: BlockT,
+    P: TransactionPool<Block = B> + 'static,
+    BE: Backend<B> + 'static,
+    C: HeaderBackend<B> + BlockBackend<B> + StorageProvider<B, BE> + 'static,
+    C: ProvideRuntimeApi<B>,
+    C::Api: StarknetRuntimeApi<B> + ConvertTransactionRuntimeApi<B>,
+    G: GenesisProvider + Send + Sync + 'static,
+    H: HasherT + Send + Sync + 'static,
+{
+    fn prepare_pending_block_with_tx_hashes(
+        &self,
+        chain_id: Felt252Wrapper,
+    ) -> Result<PendingBlockWithTxHashes, StarknetRpcApiError> {
+        let parent_hash = self.get_best_block_hash();
+        let latest_block = get_block_by_block_hash(self.client.as_ref(), parent_hash).unwrap_or_default();
+        let latest_block_header = latest_block.header();
+        let transactions = self
+            .get_pending_txs(parent_hash)?
+            .iter()
+            .map(|tx| tx.compute_hash::<H>(chain_id.0.into(), false).0)
+            .collect::<Vec<_>>();
 
+        let pending_block = PendingBlockWithTxHashes {
+            transactions,
+            l1_gas_price: latest_block_header.l1_gas_price.into(),
+            parent_hash: latest_block_header.hash::<H>().into(),
+            sequencer_address: Felt252Wrapper::from(latest_block_header.sequencer_address).into(),
+            starknet_version: latest_block_header.protocol_version.to_string(),
+            timestamp: calculate_pending_block_timestamp(),
+        };
+        Ok(pending_block)
+    }
+
+    fn prepare_pending_block_with_txs(
+        &self,
+        chain_id: Felt252Wrapper,
+    ) -> Result<PendingBlockWithTxs, StarknetRpcApiError> {
+        let parent_hash = self.get_best_block_hash();
+        let latest_block = get_block_by_block_hash(self.client.as_ref(), parent_hash).unwrap_or_default();
+        let latest_block_header = latest_block.header();
+
+        let transactions = self
+            .get_pending_txs(parent_hash)?
+            .iter()
+            .map(|tx| to_starknet_core_tx(tx.clone(), tx.compute_hash::<H>(chain_id.0.into(), false).0))
+            .collect::<Vec<_>>();
+
+        let pending_block = PendingBlockWithTxs {
+            transactions,
+            l1_gas_price: latest_block_header.l1_gas_price.into(),
+            parent_hash: latest_block_header.hash::<H>().into(),
+            sequencer_address: Felt252Wrapper::from(latest_block_header.sequencer_address).into(),
+            starknet_version: latest_block_header.protocol_version.to_string(),
+            timestamp: calculate_pending_block_timestamp(),
+        };
+        Ok(pending_block)
+    }
+
+    fn get_pending_txs(&self, latest_block: B::Hash) -> Result<Vec<mp_transactions::Transaction>, StarknetRpcApiError> {
+        // Fetch all Pending Txs from Transaction Pool
+        // Operates as RPC Call `author_pendingExtrinsics`
+        // See https://github.com/paritytech/polkadot-sdk/blob/release-polkadot-v1.6.0/substrate/client/rpc/src/author/mod.rs#L153-L155
+        // But fetches pending transactions as Vec<B::Extrinsic>
+        let pending_transactions: Vec<B::Extrinsic> = self.pool.ready().map(|tx| tx.data().clone()).collect();
+
+        // Use Runtime API to filter all Pending Txs
+        // And get only Starknet Txs (Pallet Starknet calls) as Vec<mp_transactions::Transaction>
+        self.filter_extrinsics(latest_block, pending_transactions)
+    }
+
+    async fn prepare_tx_receipt(
+        &self,
+        chain_id: Felt252Wrapper,
+        transaction_hash: FieldElement,
+        substrate_block_hash: B::Hash,
+    ) -> Result<MaybePendingTransactionReceipt, StarknetRpcApiError> {
+        let starknet_block: mp_block::Block = get_block_by_block_hash(self.client.as_ref(), substrate_block_hash)
+            .map_err(|_e| StarknetRpcApiError::BlockNotFound)?;
         let block_header = starknet_block.header();
         let block_hash = block_header.hash::<H>().into();
         let block_number = block_header.block_number;
@@ -1458,91 +1541,40 @@ where
             })?
             .ok_or(StarknetRpcApiError::BlockNotFound)?;
 
-        let chain_id = self.chain_id()?.0.into();
+        let fee_disabled = self.is_transaction_fee_disabled(substrate_block_hash)?;
 
-        let starknet_version = starknet_block.header().protocol_version;
-
-        let fee_disabled =
-            self.client.runtime_api().is_transaction_fee_disabled(substrate_block_hash).map_err(|e| {
-                error!("Failed to get check fee disabled. Substrate block hash: {substrate_block_hash}, error: {e}");
-                StarknetRpcApiError::InternalServerError
-            })?;
-
-        let block_extrinsics_len = block_extrinsics.len();
-        let transactions =
-            self.client.runtime_api().extrinsic_filter(substrate_block_hash, block_extrinsics).map_err(|e| {
-                error!("Failed to filter extrinsics. Substrate block hash: {substrate_block_hash}, error: {e}");
-                StarknetRpcApiError::InternalServerError
-            })?;
+        let transactions = self.filter_extrinsics(substrate_block_hash, block_extrinsics)?;
         let txn_hashes = self.get_cached_transaction_hashes(starknet_block.header().hash::<H>().into());
-        let mut tx_index = None;
         let mut transaction = None;
         for (index, tx) in transactions.iter().enumerate() {
             let tx_hash = self.try_txn_hash_from_cache(index, &txn_hashes, &transactions, chain_id)?;
             if tx_hash == transaction_hash.into() {
-                tx_index = Some(index);
                 transaction = Some(tx);
                 break;
             }
         }
-        if tx_index.is_none() || transaction.is_none() {
+        if transaction.is_none() {
             error!(
                 "Failed to find transaction hash in block. Substrate block hash: {substrate_block_hash}, transaction \
                  hash: {transaction_hash}"
             );
-            return Err(StarknetRpcApiError::InternalServerError.into());
+            return Err(StarknetRpcApiError::InternalServerError);
         }
-        let tx_index = tx_index.unwrap();
         let transaction = transaction.unwrap();
-        // adding the inherents count to the tx_index to get the correct index in the block
-        let tx_index = tx_index + block_extrinsics_len - transactions.len();
 
-        let events = self
-            .client
-            .runtime_api()
-            .get_events_for_tx_by_index(substrate_block_hash, tx_index as u32)
-            .map_err(|e| {
-                error!(
-                    "Failed to get events for transaction index. Substrate block hash: {substrate_block_hash}, \
-                     transaction idx: {tx_index}, error: {e}"
-                );
-                StarknetRpcApiError::InternalServerError
-            })?
-            .expect("the transaction should be present in the substrate extrinsics"); // not reachable
+        let events = self.get_events_for_tx_by_hash(substrate_block_hash, Felt252Wrapper(transaction_hash).into())?;
 
         let execution_result = {
-            let revert_error = self
-                .client
-                .runtime_api()
-                .get_tx_execution_outcome(substrate_block_hash, Felt252Wrapper(transaction_hash).into())
-                .map_err(|e| {
-                    error!(
-                        "Failed to get transaction execution outcome. Substrate block hash: {substrate_block_hash}, \
-                         transaction hash: {transaction_hash}, error: {e}"
-                    );
-                    StarknetRpcApiError::InternalServerError
-                })?;
+            let revert_error = self.get_tx_execution_outcome(substrate_block_hash, transaction_hash)?;
 
-            match revert_error {
-                None => ExecutionResult::Succeeded,
-                // This is safe because the message is a Vec<u8> build from a String
-                Some(message) => ExecutionResult::Reverted { reason: unsafe { String::from_utf8_unchecked(message) } },
-            }
+            // This is safe because the message is a Vec<u8> build from a String
+            revert_error_to_execution_result(
+                revert_error.map(|message| unsafe { String::from_utf8_unchecked(message) }),
+            )
         };
 
-        // TODO(#1291): compute message hash correctly to L1HandlerTransactionReceipt
-        let message_hash: Hash256 = Hash256::from_felt(&FieldElement::default());
-
-        fn event_conversion(event: starknet_api::transaction::Event) -> starknet_core::types::Event {
-            starknet_core::types::Event {
-                from_address: Felt252Wrapper::from(event.from_address).0,
-                keys: event.content.keys.into_iter().map(|felt| Felt252Wrapper::from(felt).0).collect(),
-                data: event.content.data.0.into_iter().map(|felt| Felt252Wrapper::from(felt).0).collect(),
-            }
-        }
-
         let events_converted: Vec<starknet_core::types::Event> =
-            events.clone().into_iter().map(event_conversion).collect();
+            events.clone().into_iter().map(starknet_api_to_starknet_core_event).collect();
 
         let actual_fee = if fee_disabled {
             FieldElement::ZERO
@@ -1560,47 +1592,34 @@ where
             events_converted.last().unwrap().data[2]
         };
 
-        let messages = self
-            .client
-            .runtime_api()
-            .get_tx_messages_to_l1(substrate_block_hash, Felt252Wrapper(transaction_hash).into())
+        let messages = self.get_tx_messages_to_l1(substrate_block_hash, transaction_hash)?;
+
+        let messages_sent = messages.into_iter().map(starknet_api_to_starknet_core_message_to_l1).collect();
+
+        // TODO
+        // Is any better way to get execution resources of processed tx?
+        let skip_validate = true;
+        // let parent_block_hash = Felt252Wrapper::from().into();
+        let parent_block_hash = self
+            .substrate_block_hash_from_starknet_block(BlockId::Hash(block_header.parent_block_hash.into()))
             .map_err(|e| {
-                error!("'{e}'");
-                StarknetRpcApiError::InternalServerError
+                error!("Parent Block not found: {e}");
+                StarknetRpcApiError::BlockNotFound
             })?;
+        let simulation = self.simulate_tx(parent_block_hash, transaction.clone(), skip_validate, fee_disabled)?;
+        let execution_resources = actual_resources_to_execution_resources(simulation.actual_resources);
 
-        fn message_conversion(message: starknet_api::transaction::MessageToL1) -> starknet_core::types::MsgToL1 {
-            let mut to_address = [0u8; 32];
-            to_address[12..32].copy_from_slice(message.to_address.0.as_bytes());
-            starknet_core::types::MsgToL1 {
-                from_address: Felt252Wrapper::from(message.from_address).0,
-                to_address: FieldElement::from_bytes_be(&to_address).unwrap(),
-                payload: message.payload.0.into_iter().map(|felt| Felt252Wrapper::from(felt).0).collect(),
-            }
-        }
-
-        // TODO: use actual execution ressources
         let receipt = match transaction {
-            mp_transactions::Transaction::Declare(_) => TransactionReceipt::Declare(DeclareTransactionReceipt {
+            mp_transactions::Transaction::Declare(_, _) => TransactionReceipt::Declare(DeclareTransactionReceipt {
                 transaction_hash,
                 actual_fee,
                 finality_status: TransactionFinalityStatus::AcceptedOnL2,
                 block_hash,
                 block_number,
-                messages_sent: messages.into_iter().map(message_conversion).collect(),
+                messages_sent,
                 events: events_converted,
                 execution_result,
-                execution_resources: ExecutionResources {
-                    steps: 0,
-                    memory_holes: None,
-                    range_check_builtin_applications: 0,
-                    pedersen_builtin_applications: 0,
-                    poseidon_builtin_applications: 0,
-                    ec_op_builtin_applications: 0,
-                    ecdsa_builtin_applications: 0,
-                    bitwise_builtin_applications: 0,
-                    keccak_builtin_applications: 0,
-                },
+                execution_resources,
             }),
             mp_transactions::Transaction::DeployAccount(tx) => {
                 TransactionReceipt::DeployAccount(DeployAccountTransactionReceipt {
@@ -1609,21 +1628,11 @@ where
                     finality_status: TransactionFinalityStatus::AcceptedOnL2,
                     block_hash,
                     block_number,
-                    messages_sent: messages.into_iter().map(message_conversion).collect(),
+                    messages_sent,
                     events: events_converted,
                     contract_address: tx.get_account_address(),
                     execution_result,
-                    execution_resources: ExecutionResources {
-                        steps: 0,
-                        memory_holes: None,
-                        range_check_builtin_applications: 0,
-                        pedersen_builtin_applications: 0,
-                        poseidon_builtin_applications: 0,
-                        ec_op_builtin_applications: 0,
-                        ecdsa_builtin_applications: 0,
-                        bitwise_builtin_applications: 0,
-                        keccak_builtin_applications: 0,
-                    },
+                    execution_resources,
                 })
             }
             mp_transactions::Transaction::Invoke(_) => TransactionReceipt::Invoke(InvokeTransactionReceipt {
@@ -1632,46 +1641,150 @@ where
                 finality_status: TransactionFinalityStatus::AcceptedOnL2,
                 block_hash,
                 block_number,
-                messages_sent: messages.into_iter().map(message_conversion).collect(),
+                messages_sent,
                 events: events_converted,
                 execution_result,
-                execution_resources: ExecutionResources {
-                    steps: 0,
-                    memory_holes: None,
-                    range_check_builtin_applications: 0,
-                    pedersen_builtin_applications: 0,
-                    poseidon_builtin_applications: 0,
-                    ec_op_builtin_applications: 0,
-                    ecdsa_builtin_applications: 0,
-                    bitwise_builtin_applications: 0,
-                    keccak_builtin_applications: 0,
-                },
+                execution_resources,
             }),
-            mp_transactions::Transaction::L1Handler(_) => TransactionReceipt::L1Handler(L1HandlerTransactionReceipt {
-                message_hash,
-                transaction_hash,
-                actual_fee,
-                finality_status: TransactionFinalityStatus::AcceptedOnL2,
-                block_hash,
-                block_number,
-                messages_sent: messages.into_iter().map(message_conversion).collect(),
-                events: events_converted,
-                execution_result,
-                execution_resources: ExecutionResources {
-                    steps: 0,
-                    memory_holes: None,
-                    range_check_builtin_applications: 0,
-                    pedersen_builtin_applications: 0,
-                    poseidon_builtin_applications: 0,
-                    ec_op_builtin_applications: 0,
-                    ecdsa_builtin_applications: 0,
-                    bitwise_builtin_applications: 0,
-                    keccak_builtin_applications: 0,
-                },
-            }),
+            mp_transactions::Transaction::L1Handler(ref tx) => {
+                TransactionReceipt::L1Handler(L1HandlerTransactionReceipt {
+                    message_hash: Hash256::from_felt(&tx.compute_hash::<H>(chain_id, false).0),
+                    transaction_hash,
+                    actual_fee,
+                    finality_status: TransactionFinalityStatus::AcceptedOnL2,
+                    block_hash,
+                    block_number,
+                    messages_sent,
+                    events: events_converted,
+                    execution_result,
+                    execution_resources,
+                })
+            }
         };
 
         Ok(MaybePendingTransactionReceipt::Receipt(receipt))
+    }
+
+    fn get_events_for_tx_by_hash(
+        &self,
+        substrate_block_hash: B::Hash,
+        tx_hash: TransactionHash,
+    ) -> Result<Vec<starknet_api::transaction::Event>, StarknetRpcApiError> {
+        let events = self.do_get_events_for_tx_by_hash(substrate_block_hash, tx_hash)?;
+        Ok(events)
+    }
+
+    fn get_tx_execution_outcome(
+        &self,
+        substrate_block_hash: B::Hash,
+        transaction_hash: FieldElement,
+    ) -> Result<Option<Vec<u8>>, StarknetRpcApiError> {
+        self.do_get_tx_execution_outcome(substrate_block_hash, Felt252Wrapper(transaction_hash).into())
+    }
+
+    fn find_pending_tx(
+        &self,
+        chain_id: Felt252Wrapper,
+        tx_hash: FieldElement,
+    ) -> Result<Option<mp_transactions::Transaction>, StarknetRpcApiError> {
+        let latest_block = self.get_best_block_hash();
+
+        let pending_tx = self
+            .get_pending_txs(latest_block)?
+            .iter()
+            .find(|&tx| tx.compute_hash::<H>(chain_id.0.into(), false).0 == tx_hash)
+            .cloned();
+
+        Ok(pending_tx)
+    }
+
+    async fn get_pending_transaction_receipt(
+        &self,
+        chain_id: Felt252Wrapper,
+        transaction_hash: FieldElement,
+    ) -> Result<MaybePendingTransactionReceipt, StarknetRpcApiError> {
+        let pending_tx =
+            self.find_pending_tx(chain_id, transaction_hash)?.ok_or(StarknetRpcApiError::TxnHashNotFound)?;
+
+        // TODO -- no way of getting messages sent to L1 for the tx
+        let messages_sent = Vec::new();
+        // TODO -- no  way of getting events for the tx
+        let events = Vec::new();
+
+        // TODO -- should Tx be simulated with `skip_validate`?
+        let skip_validate = true;
+        let skip_fee_charge = self.is_transaction_fee_disabled(self.get_best_block_hash())?;
+        let simulation =
+            self.simulate_tx(self.get_best_block_hash(), pending_tx.clone(), skip_validate, skip_fee_charge)?;
+        let actual_fee = simulation.actual_fee.0.into();
+        let execution_result = revert_error_to_execution_result(simulation.revert_error);
+        let execution_resources = actual_resources_to_execution_resources(simulation.actual_resources);
+
+        let receipt = match pending_tx {
+            mp_transactions::Transaction::Declare(_tx, _contract_class) => {
+                let receipt = PendingDeclareTransactionReceipt {
+                    transaction_hash,
+                    actual_fee,
+                    messages_sent,
+                    events,
+                    execution_resources,
+                    execution_result,
+                };
+                PendingTransactionReceipt::Declare(receipt)
+            }
+            mp_transactions::Transaction::DeployAccount(ref tx) => {
+                let contract_address = tx.get_account_address();
+                let receipt = PendingDeployAccountTransactionReceipt {
+                    transaction_hash,
+                    actual_fee,
+                    messages_sent,
+                    events,
+                    execution_resources,
+                    execution_result,
+                    contract_address,
+                };
+                PendingTransactionReceipt::DeployAccount(receipt)
+            }
+            mp_transactions::Transaction::Invoke(_tx) => {
+                let receipt = PendingInvokeTransactionReceipt {
+                    transaction_hash,
+                    actual_fee,
+                    messages_sent,
+                    events,
+                    execution_resources,
+                    execution_result,
+                };
+                PendingTransactionReceipt::Invoke(receipt)
+            }
+            mp_transactions::Transaction::L1Handler(ref tx) => {
+                let receipt = PendingL1HandlerTransactionReceipt {
+                    message_hash: Hash256::from_felt(&tx.compute_hash::<H>(chain_id, false).0),
+                    transaction_hash,
+                    actual_fee,
+                    messages_sent,
+                    events,
+                    execution_resources,
+                    execution_result,
+                };
+                PendingTransactionReceipt::L1Handler(receipt)
+            }
+        };
+
+        Ok(MaybePendingTransactionReceipt::PendingReceipt(receipt))
+    }
+
+    fn convert_error<T>(
+        &self,
+        best_block_hash: <B as BlockT>::Hash,
+        call_result: Result<T, DispatchError>,
+    ) -> Result<T, StarknetRpcApiError> {
+        match call_result {
+            Ok(val) => Ok(val),
+            Err(e) => {
+                let starknet_error = self.convert_dispatch_error(best_block_hash, e)?;
+                Err(starknet_error.into())
+            }
+        }
     }
 }
 
@@ -1694,39 +1807,56 @@ where
     })
 }
 
-async fn convert_tx_to_extrinsic<C, B>(
-    client: Arc<C>,
-    best_block_hash: <B as BlockT>::Hash,
-    transaction: UserTransaction,
-) -> Result<<B as BlockT>::Extrinsic, StarknetRpcApiError>
-where
-    B: BlockT,
-    C: ProvideRuntimeApi<B>,
-    C::Api: StarknetRuntimeApi<B> + ConvertTransactionRuntimeApi<B>,
-{
-    let extrinsic = client.runtime_api().convert_transaction(best_block_hash, transaction).map_err(|e| {
-        error!("Failed to convert transaction: {:?}", e);
-        StarknetRpcApiError::InternalServerError
-    })?;
-
-    Ok(extrinsic)
+/// The current timestamp in seconds.
+fn calculate_pending_block_timestamp() -> u64 {
+    let timestamp_in_millisecond = sp_timestamp::InherentDataProvider::from_system_time().as_millis();
+    timestamp_in_millisecond / 1000
 }
 
-fn convert_error<C, B, T>(
-    client: Arc<C>,
-    best_block_hash: <B as BlockT>::Hash,
-    call_result: Result<T, DispatchError>,
-) -> Result<T, StarknetRpcApiError>
-where
-    B: BlockT,
-    C: ProvideRuntimeApi<B>,
-    C::Api: StarknetRuntimeApi<B> + ConvertTransactionRuntimeApi<B>,
-{
-    match call_result {
-        Ok(val) => Ok(val),
-        Err(e) => match client.runtime_api().convert_error(best_block_hash, e) {
-            Ok(starknet_error) => Err(starknet_error.into()),
-            Err(_) => Err(StarknetRpcApiError::InternalServerError),
-        },
+fn is_pending_block(block_id: BlockId) -> bool {
+    block_id == BlockId::Tag(BlockTag::Pending)
+}
+
+fn starknet_api_to_starknet_core_event(event: starknet_api::transaction::Event) -> starknet_core::types::Event {
+    starknet_core::types::Event {
+        from_address: Felt252Wrapper::from(event.from_address).0,
+        keys: event.content.keys.into_iter().map(|felt| Felt252Wrapper::from(felt).0).collect(),
+        data: event.content.data.0.into_iter().map(|felt| Felt252Wrapper::from(felt).0).collect(),
+    }
+}
+
+fn starknet_api_to_starknet_core_message_to_l1(
+    message: starknet_api::transaction::MessageToL1,
+) -> starknet_core::types::MsgToL1 {
+    let mut to_address = [0u8; 32];
+    to_address[12..32].copy_from_slice(message.to_address.0.as_bytes());
+    starknet_core::types::MsgToL1 {
+        from_address: Felt252Wrapper::from(message.from_address).0,
+        to_address: FieldElement::from_bytes_be(&to_address).unwrap(),
+        payload: message.payload.0.into_iter().map(|felt| Felt252Wrapper::from(felt).0).collect(),
+    }
+}
+
+fn revert_error_to_execution_result(revert_error: Option<String>) -> ExecutionResult {
+    match revert_error {
+        None => ExecutionResult::Succeeded,
+        Some(message) => ExecutionResult::Reverted { reason: message },
+    }
+}
+
+fn actual_resources_to_execution_resources(resources: ResourcesMapping) -> ExecutionResources {
+    let resources = resources.0.into_iter().map(|(k, v)| (k.to_lowercase(), v)).collect::<HashMap<String, u64>>();
+    // Based on `VM_RESOURCE_FEE_COSTS`
+    // in crates/primitives/fee/src/lib.rs
+    ExecutionResources {
+        steps: *resources.get("n_steps").unwrap_or(&0),
+        memory_holes: resources.get("memory_holes").copied(),
+        range_check_builtin_applications: *resources.get("range_check_builtin").unwrap_or(&0),
+        pedersen_builtin_applications: *resources.get("pedersen_builtin").unwrap_or(&0),
+        poseidon_builtin_applications: *resources.get("poseidon_builtin").unwrap_or(&0),
+        ec_op_builtin_applications: *resources.get("ec_op_builtin").unwrap_or(&0),
+        ecdsa_builtin_applications: *resources.get("ecdsa_builtin").unwrap_or(&0),
+        bitwise_builtin_applications: *resources.get("bitwise_builtin").unwrap_or(&0),
+        keccak_builtin_applications: *resources.get("keccak_builtin").unwrap_or(&0),
     }
 }